common: move set_before_first_get_setting_t from cpu/x64/cpu_isa_traits.cpp to common/utils.hpp It is used by not only x64 but also various architecture.
